requirements.txt
# Requirements for DocTR Payslip Extractor

# Core OCR and Document Processing
torch>=1.9.0
torchvision>=0.10.0
python-doctr[torch]>=0.6.0

# Web Interface
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
Pillow>=8.3.0

# File Handling
pathlib2>=2.3.0
python-magic>=0.4.24

# Utilities
regex>=2022.7.9
python-dateutil>=2.8.0
tqdm>=4.64.0

# Export and Visualization
matplotlib>=3.5.0
plotly>=5.11.0

# Optional: For better PDF handling
PyPDF2>=3.0.0
pdf2image>=1.16.0

# Development and Testing (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
